Fix issues and improve tasks
