+++ .gitignore
@@ -18,6 +18,7 @@ ignition/deployments/chain-31337
 
 node_modules
 .env
+.idea
 
 # Hardhat files
 /cache
@@ -32,4 +33,4 @@ node_modules
 /coverage.json
 
 # Hardhat Ignition default folder for deployments against a local node
-ignition/deployments/chain-31337
+ignition/deployments/
